Verificaciones y checkAuth Anuncios

+++ controllers/Announcement.controller.js
@@ -11,6 +11,9 @@ export const getAllAnnouncements = async (req, res) => {
 
 export const createAnnouncement = async (req, res) => {
  try {
+   if (!req.body.name || !req.body.description || !req.body.date || !req.body.image) {
+      return res.status(400).json({ message: 'Por favor, ingrese todos los campos requeridos' });
+    }
     const announcement = new Announcement({
       name: req.body.name,
       description: req.body.description,
@@ -29,6 +32,9 @@ export const createAnnouncement = async (req, res) => {
 export const getAnnouncement = async (req, res) => {
  try {
     const announcement = await Announcement.findById(req.params.id);
+    if(!announcement){
+      return res.status(404).json({ message: 'Anuncio no encontrado' });
+    }
     res.status(200).json(announcement);
  } catch (error) {
     res.status(400).json({ message: error.message });
@@ -37,6 +43,13 @@ export const getAnnouncement = async (req, res) => {
 
 export const updateAnnouncement = async (req, res) => {
  try {
+   if (!req.body.name || !req.body.description || !req.body.date || !req.body.image) {
+      return res.status(400).json({ message: 'Por favor, ingrese todos los campos requeridos' });
+    }
+    const AnnouncementExist = await Announcement.findById(req.params.id);
+    if(!Announcement) {
+      return res.status(404).json({ message: 'Anuncio no encontrado' });
+    }
     const announcement = await Announcement.findByIdAndUpdate(
       req.params.id,
       {
@@ -56,6 +69,10 @@ export const updateAnnouncement = async (req, res) => {
 
 export const deleteAnnouncement = async (req, res) => {
  try {
+   const Announcement = await Announcement.findById(req.params.id);
+   if(!Announcement) {
+     return res.status(404).json({ message: 'Anuncio no encontrado' });
+   }
     await Announcement.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: 'Announcement deleted successfully' });
  } catch (error) {
